// src/shared/services/api/models/courseOverviewDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Learning Assessment Portal API
 * OpenAPI spec version: v1
 */
import type { CourseOverviewMetaTopicDto } from './courseOverviewMetaTopicDto';
import type { RefTermDto } from './refTermDto';
import type { UserSummaryDto } from './userSummaryDto';

export interface CourseOverviewDto {
  id?: string;
  /** @nullable */
  title?: string | null;
  category?: RefTermDto;
  difficulty_level?: RefTermDto;
  duration_minute?: number;
  overall_rating?: number;
  /** @nullable */
  thumbnail_img?: string | null;
  is_drafted?: boolean;
  /** @nullable */
  description?: string | null;
  sub_category?: RefTermDto;
  created_by_user?: UserSummaryDto;
  date_created?: string;
  date_updated?: string;
  enrollment_count?: number;
  /** @nullable */
  assessment_title?: string | null;
  total_mark?: number;
  passing_mark?: number;
  /** @nullable */
  topic?: CourseOverviewMetaTopicDto[] | null;
}