// src/shared/services/api/models/courseOverviewMetaTopicDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Learning Assessment Portal API
 * OpenAPI spec version: v1
 */
import type { CourseOverviewContentDto } from './courseOverviewContentDto';

export interface CourseOverviewMetaTopicDto {
  id?: string;
  /** @nullable */
  name?: string | null;
  meta_sequence_order?: number;
  meta_duration_minute?: number;
  /** @nullable */
  contents?: CourseOverviewContentDto[] | null;
}