// src/shared/services/api/models/courseContentProgressDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Learning Assessment Portal API
 * OpenAPI spec version: v1
 */
import type { RefTermDto } from './refTermDto';

export interface CourseContentProgressDto {
  id?: string;
  meta_topic_id?: string;
  /** @nullable */
  title?: string | null;
  content_type?: RefTermDto;
  /** @nullable */
  video_url?: string | null;
  /** @nullable */
  pdf_file_path?: string | null;
  sequence_order?: number;
  is_completed?: boolean;
  /** @nullable */
  completed_on?: string | null;
}